# supporting constexpr values/types in variant