--- a/include/glaze/json/read.hpp
+++ b/include/glaze/json/read.hpp
@@ -1687,15 +1687,15 @@ namespace glz
       struct variant_types<std::variant<Ts...>>
       {
          // TODO this way of filtering types is compile time intensive.
-         using bool_types = decltype(std::tuple_cat(std::conditional_t<bool_t<Ts>, std::tuple<Ts>, std::tuple<>>{}...));
+         using bool_types = decltype(std::tuple_cat(std::conditional_t<bool_t<remove_meta_wrapper_t<Ts>>, std::tuple<Ts>, std::tuple<>>{}...));
          using number_types = decltype(std::tuple_cat(
             std::conditional_t<num_t<remove_meta_wrapper_t<Ts>>, std::tuple<Ts>, std::tuple<>>{}...));
-         using string_types = decltype(std::tuple_cat(std::conditional_t < str_t<Ts> || glaze_enum_t<Ts>,
+         using string_types = decltype(std::tuple_cat(std::conditional_t < str_t<remove_meta_wrapper_t<Ts>> || glaze_enum_t<Ts>,
                                                       std::tuple<Ts>, std::tuple < >> {}...));
          using object_types =
             decltype(std::tuple_cat(std::conditional_t < readable_map_t<Ts> || writable_map_t<Ts> || glaze_object_t<Ts>,
                                     std::tuple<Ts>, std::tuple < >> {}...));
-         using array_types = decltype(std::tuple_cat(std::conditional_t < array_t<Ts> || glaze_array_t<Ts>,
+         using array_types = decltype(std::tuple_cat(std::conditional_t < array_t<remove_meta_wrapper_t<Ts>> || glaze_array_t<Ts>,
                                                      std::tuple<Ts>, std::tuple < >> {}...));
          using nullable_types =
             decltype(std::tuple_cat(std::conditional_t<null_t<Ts>, std::tuple<Ts>, std::tuple<>>{}...));
@@ -1712,6 +1712,53 @@ namespace glz
             decltype(std::tuple_cat(std::conditional_t<glaze_const_value_t<Ts>, std::tuple<Ts>, std::tuple<>>{}...));
          using glaze_non_const_types =
             decltype(std::tuple_cat(std::conditional_t<!glaze_const_value_t<Ts>, std::tuple<Ts>, std::tuple<>>{}...));
+      };
+
+      template <typename tuple_types_t>
+      struct process_arithmetic_boolean_string_or_array
+      {
+         template <auto Options>
+         GLZ_FLATTEN static void op(auto&& value, is_context auto&& ctx, auto&& it, auto&& end)
+         {
+            if constexpr (std::tuple_size_v<tuple_types_t> < 1) {
+               ctx.error = error_code::no_matching_variant_type;
+            }
+            else {
+               using const_glaze_types = typename tuple_types<tuple_types_t>::glaze_const_types;
+               bool found_match{};
+               for_each<std::tuple_size_v<const_glaze_types>>([&]([[maybe_unused]] auto I) mutable {
+                  if (found_match) {
+                     return;
+                  }
+                  using V = std::tuple_element_t<I, const_glaze_types>;
+                  // run time substitute to compare to const value
+                  std::remove_const_t<std::remove_pointer_t<std::remove_const_t<meta_wrapper_t<V>>>> substitute{};
+                  auto copy_it{it};
+                  read<json>::op<ws_handled<Options>()>(substitute, ctx, it, end);
+                  static constexpr auto const_value{*meta_wrapper_v<V>};
+                  if (substitute == const_value) {
+                     found_match = true;
+                     if (!std::holds_alternative<V>(value)) value = V{};
+                  }
+                  else {
+                     it = copy_it;
+                  }
+               });
+               if (found_match) {
+                  return;
+               }
+
+               using non_const_types = typename tuple_types<tuple_types_t>::glaze_non_const_types;
+               if constexpr (std::tuple_size_v<non_const_types> > 0) {
+                  using V = std::tuple_element_t<0, non_const_types>;
+                  if (!std::holds_alternative<V>(value)) value = V{};
+                  read<json>::op<ws_handled<Options>()>(std::get<V>(value), ctx, it, end);
+               }
+               else {
+                  ctx.error = error_code::no_matching_variant_type;
+               }
+            }
+         }
       };
 
       template <is_variant T>
@@ -1895,38 +1942,17 @@ namespace glz
                   break;
                case '[':
                   using array_types = typename variant_types<T>::array_types;
-                  if constexpr (std::tuple_size_v<array_types> < 1) {
-                     ctx.error = error_code::no_matching_variant_type;
-                  }
-                  else {
-                     using V = std::tuple_element_t<0, array_types>;
-                     if (!std::holds_alternative<V>(value)) value = V{};
-                     read<json>::op<ws_handled<Opts>()>(std::get<V>(value), ctx, it, end);
-                  }
+                  process_arithmetic_boolean_string_or_array<array_types>::template op<Opts>(value, ctx, it, end);
                   break;
                case '"': {
                   using string_types = typename variant_types<T>::string_types;
-                  if constexpr (std::tuple_size_v<string_types> < 1) {
-                     ctx.error = error_code::no_matching_variant_type;
-                  }
-                  else {
-                     using V = std::tuple_element_t<0, string_types>;
-                     if (!std::holds_alternative<V>(value)) value = V{};
-                     read<json>::op<ws_handled<Opts>()>(std::get<V>(value), ctx, it, end);
-                  }
+                  process_arithmetic_boolean_string_or_array<string_types>::template op<Opts>(value, ctx, it, end);
                   break;
                }
                case 't':
                case 'f': {
                   using bool_types = typename variant_types<T>::bool_types;
-                  if constexpr (std::tuple_size_v<bool_types> < 1) {
-                     ctx.error = error_code::no_matching_variant_type;
-                  }
-                  else {
-                     using V = std::tuple_element_t<0, bool_types>;
-                     if (!std::holds_alternative<V>(value)) value = V{};
-                     read<json>::op<ws_handled<Opts>()>(std::get<V>(value), ctx, it, end);
-                  }
+                  process_arithmetic_boolean_string_or_array<bool_types>::template op<Opts>(value, ctx, it, end);
                   break;
                }
                case 'n':
@@ -1943,40 +1969,7 @@ namespace glz
                default: {
                   // Not bool, string, object, or array so must be number or null
                   using number_types = typename variant_types<T>::number_types;
-                  if constexpr (std::tuple_size_v<number_types> < 1) {
-                     ctx.error = error_code::no_matching_variant_type;
-                  }
-                  else {
-                     using const_glaze_types = typename tuple_types<number_types>::glaze_const_types;
-                     bool found_match{};
-                     for_each<std::tuple_size_v<const_glaze_types>>([&]([[maybe_unused]] auto I) mutable  {
-                        if (found_match) {
-                           return;
-                        }
-                        using V = std::tuple_element_t<I, const_glaze_types>;
-                        // run time substitute to compare to const value
-                        std::remove_const_t<std::remove_pointer_t<std::remove_const_t<meta_wrapper_t<V>>>> substitute{};
-                        auto copy_it{it};
-                        read<json>::op<ws_handled<Opts>()>(substitute, ctx, it, end);
-                        static constexpr auto const_value{*meta_wrapper_v<V>};
-                        if (substitute == const_value) {
-                           found_match = true;
-                           if (!std::holds_alternative<V>(value)) value = V{};
-                        }
-                        else {
-                           it = copy_it;
-                        }
-                     });
-                     using other_types = typename tuple_types<number_types>::glaze_non_const_types;
-                     if constexpr (std::tuple_size_v<other_types> > 0) {
-                        using V = std::tuple_element_t<0, other_types>;
-                        if (!std::holds_alternative<V>(value)) value = V{};
-                        read<json>::op<ws_handled<Opts>()>(std::get<V>(value), ctx, it, end);
-                     }
-                     else if (!found_match) {
-                        ctx.error = error_code::no_matching_variant_type;
-                     }
-                  }
+                  process_arithmetic_boolean_string_or_array<number_types>::template op<Opts>(value, ctx, it, end);
                }
                }
             }
@@ -2097,36 +2090,6 @@ namespace glz
             }
          }
       };
-
-      // template <glaze_const_value_t T>
-      // struct from_json<T>
-      // {
-      //    // First attempt, This fails because you will read and increment `it` for each none_match so second attempt is
-      //    // indexing bogus memory etc. Would be nice if this could be here
-      //    template <auto Options>
-      //    GLZ_FLATTEN static void op([[maybe_unused]] auto&& value, [[maybe_unused]] is_context auto&& ctx,
-      //                               [[maybe_unused]] auto&& it, [[maybe_unused]] auto&& end) noexcept
-      //    {
-      //       auto copy_it{it};
-      //       std::remove_const_t<std::remove_pointer_t<std::remove_const_t<meta_wrapper_t<T>>>> substitute{};
-      //       read<json>::op<Options>(substitute, ctx, it, end);
-      //
-      //       // from_json<decltype(substitute)>::template op<Options>(substitute, ctx, it, end);
-      //       static constexpr auto const_value{*meta_wrapper_v<T>};
-      //       if (substitute == const_value) {
-      //          if (ctx.error == error_code::no_matching_variant_type) {
-      //             // clear error because we found a match
-      //             ctx.error = error_code::none;
-      //          }
-      //       }
-      //       else {
-      //          // reset it to its previous form because we did not read an actual value
-      //          it = copy_it;
-      //          ctx.error = error_code::no_matching_variant_type;
-      //       }
-      //    }
-      // };
-
    } // namespace detail
 
    template <class Buffer>