Added support for handling reads of constexpr struct values in variants.

Introduced type handling of const types for glaze value, and split the const types from the non const types during the parse of variant types.