@jbbjarnason 2023-11-21
very unofficial compilation time benchmarking

before
```bash
jonb@jonb-desktop ~/P/glaze (support-direct-constexpr-read)> time cmake --build build --target json_test
[0/2] Re-checking globbed directories...
[2/2] Linking CXX executable tests/json_test/json_test

________________________________________________________
Executed in   43.86 secs    fish           external
   usr time   40.95 secs  246.00 micros   40.95 secs
   sys time    2.76 secs  351.00 micros    2.76 secs
```

after
```bash
jonb@jonb-desktop ~/P/glaze (support-direct-constexpr-read) [1]> time cmake --build build --target json_test
[0/2] Re-checking globbed directories...
[2/2] Linking CXX executable tests/json_test/json_test

________________________________________________________
Executed in   42.71 secs    fish           external
   usr time   40.01 secs  233.00 micros   40.01 secs
   sys time    2.58 secs  345.00 micros    2.58 secs
```

Does not show any major difference but it's a bit better