@stephenberry 2023-11-21
Thanks for the benchmark and making those changes. I guess there's not a significant difference in this context, but I've seen in make a big difference elsewhere.